chore(media module): rename object detail module to media
