+++ scripts/path-aliases.js
@@ -1,3 +1,12 @@
-const pathAliases = ['auth', 'config', 'home', 'i18n', 'navigation', 'reading-room', 'shared'];
+const pathAliases = [
+	'auth',
+	'config',
+	'home',
+	'i18n',
+	'media',
+	'navigation',
+	'reading-room',
+	'shared',
+];
 
 module.exports = pathAliases;
