+++ src/modules/media/components/Metadata/Metadata.module.scss
