+++ src/modules/media/components/ObjectPlaceholder/ObjectPlaceholder.module.scss
