+++ src/modules/media/components/ObjectPlaceholder/ObjectPlaceholder.stories.tsx
