+++ src/modules/media/components/ObjectPlaceholder/__mocks__/object-placeholder.ts
