+++ tsconfig.json
@@ -6,6 +6,7 @@
       "@config/*": ["modules/config/*"],
       "@home/*": ["modules/home/*"],
       "@i18n/*": ["modules/i18n/*"],
+      "@media/*": ["modules/media/*"],
       "@navigation/*": ["modules/navigation/*"],
       "@reading-room/*": ["modules/reading-room/*"],
       "@shared/*": ["modules/shared/*"]
